Initial inheritance relation of Object/Player done
This commit is the first step in making the Object
class the base class for all future in game entities
that have movement (velocity,accel) and even mass.
Objects do not have sprites. In order to create an
object with a sprite, derive a class from object and
compose the new class with the new Sprite class. The
new sprite class handles everything from rotation
creation to rotational animation and much more in
the future (auto collison box creation, deformation,
 etc.).This allows easy creation of new on screen
objects/entities and their behaviors. Also, with
the movement and physics effects now separated from
the animation, Objects can be used to create
interesting effects. NOTE: This commit does not
introduce any new demos or behvaior that takes advantage
of the new back end refactoring. Everything operates as
before (except a minore bug fix with the "w" key).