Added background support!
No more plain black backgrounds.
This is the first implementation of
backgrounds with alpha layers (PNG).
Nex step will be to create a layered
background that makes use of Alpha
channel to give the first sense of
depth that this engine has seen. Later
add support for animated backgrounds.
This also introduces a problem that I've
been aware of for some time. The pixel
dimming technique in that past has simply
been to slowly darken the pixels to black.
This will no longer work nicely (look good)
now that backgrounds dont have to be
black.