Disabled Object orbit. Collisions!
I had to disable gravitational effects
between objects as it was causing major
stability issues. The behavior is odd,
considerably undefined. Sometimes things
work, sometimes they dont. This issue will
have to be fixed as it is a wanted and
expected feature. Additionally, collision
detections has made its first, very rough,
appearance in the form of hit circles. It
looks like this method could work pretty
well, but some major overhauls will be
needed in the way the physics is
implemented. Should be fun.