Refined Object to Player polymorphism
The parent class "Object" now contains
all the methods and fields needed to get
something moving and rotating on the screen.
This leaves the player class to focus on
more player oriented attributes, such as
team, weapons, health, fuel, etc. This
will be further demostrated with the
Missile class. Additionally, the previous
commit included a fix for a major program flaw that
had gone unnoticed for quite a few commits.
This flaw is also the possible culprit for
the undefined behavior that was being experienced.
The error was in the use of the pthread read/write
lock, more specifically, failing to initialize
the object_rw_lock in the "RuPPAT" class.