Pixels with alpha channel added
A fair amount of changes were required
to acheive the placement of semi-transparent
(alpha channel) pixels. By using the
pre_surface field with alpha enabled and
blitting that to the main screen RuPPAT
can now place individual pixels of unique
opacity levels. This means the dimming is
now truly fading out, not simply fading to
black.