Start of config file use: parsing
Beginning steps towards full implementation
of a configure script/file. This should make
it much easier to setup different demos/games.
Note that the info from the configure is not
used for *anything* at this point. Still,
this commit requires that "config" exists
in the directory. A sample config is in the
branch and tracked. Next parsing feature is
comments within the config.