Removed RuPPAT Globals; Ifdef on Player
A lot of variables (mainly lists and locks)
were forcibly implemented as globals within
RuPPAT.cpp due to pThreads complications.
With the implementation of STD11 threads,
these variables can be moved to where they
belong, in the private section of the RuPPAT
object. This makes for much cleaner and much
more controlled code.

In doing this, Player include was added to
RuPPAT.h (previously in RuPPAT.cpp) and an
ifdef was added to Player.h to prevent
redefinitions