Moving frame (or view) implemented to follow ship
This commit implements an early version of a focused
camera. This way, a large picture can be loaded as the
background, and the much smaller resolution frame will
adjust to keep the ship in focus. Everything still
renders to a surface, only what is shown is adjusted.
Future improvements in efficiency can change this to
only blit or render what is visible to the focused
ship.