Added pixel removal to Sprite class
This is the first step to deformation,
destruction, and general manipulation
of sprites in realtime. It takes a bit
of geometry to allow the engine to
avoid re-rotating a newly deformed
base sprite. Instead, the manipulated
pixels of the base rotation can be
transformed and applied to each rotation.
This should be a much faster process than
altering the base sprite and then parsing
the rotations again using rotozoom.