Initial decoupling of Physics from objects
All gravitational physics depends on Object
class in a more refined way. Objects can be
"acted on" by a source, but they can also
"act on" a destination. This can be used in
the future to decrease loop iterations. It
also permits Pixels and Objects to be handled
nearly the same; the base struct of the pixel
(Entity_desc) is used for its physics now.

The Physics handler in Object is a virtual
function, allowing it to be overidden when it's
derived, thus different Classes (player, weapon,
etc.) derived from Object can behave completely
differently with regards to physics (this will be
the method of implementation for other things, such
as collisions).