Lots of backend changes; Missiles emergin
To better facilitate expansion of the object
class and its derived classes, quite a few
changes have been made. Most importantly,
the physics is now handled by the Object objects
themselves. In fact, these PhysicsHandler functions
are Virtual, allowing them to be easily overidden
in the derived classes. This also allows for simple
execution of the handler function; inherent in the
virtual function standard is the execution of the
most derived version of the function for that object
of any base/derivation level. Very useful.

Now, the Player class, derived from the Object class,
overides the physics handler to include calculations
for any launched Missiles. Which bring us to the next
major addition: *Missiles*. Any object-like weapon,
affected by physics and collidable, should be based
on Missile. The Player class is now composed of a
vector of missiles and fire missiles. The object takes
care of updating the missile once it has been launched.
Using the space bar now files the "White Rocket", which
means it does nothing more than spawn it and let it sit.
Though it is affected by gravity. A few things to clean
up still (alpha seems to be broken on the missile sprite)
but the concept is there.

Quite a few additions were made to help this along,
mainly implementing rule-of-3 things on the Sprite
class to allow better data movement. Also, the
PhysFuncs has now been separated out into prototypes
and definitions as it is being included into far too
many classes for a simple definition to suffice. The
PhysFunction namespace is quite a mess at the moment,
but that should change once the Missiles are better
implemented and my focuses shifts to physics revamping
and collision detection.