Fixed minor Missile bug; Added orange blob
Fixed a minor bug with the Missile demo implementation,
where the x coordinate value was being assigned to both
x and y of the new missile. Also, added an orange blob
sprite as a place holder for the missile, this way
early development of collision detection and rotational
physics can be simplified for the meantime.