Default Missile more configurable
The configuration file is now setup to be
able to define more parameters for a players
starting, or default, missile. This includes
things like the amount of missiles the player
has, the damage, and the initial velocity magnitude.
Most of these still dont affect the actual missile
during gameplay yet, but that should change soon.