Timing support partially implemented
Timing on Missiles is partially implemented;
this needs more refinement before moved any further.
The way it is implemented with Missiles will dictate
how it is implemented in future composed entities.
The implementation needs to support appropriate
destruction of each composed class, or whatever
happens when the time is reached. Furthermore, it
may be best to use the games time rather than SDL ticks
as lag or slow computers would greatly alter the
nuances of these timers.

It may also me better to use some kind of signal rather
than standard variables and polling.