Timing on Player Missiles implemented
Missiles owned by player now have a working
lifespan component. Once the lifespan is up,
the GameDestroy function is called on the missile,
and physics is no longer calculated for that missile.
The GameDestroy function is a virtual function
from Object and is currently undefined. This means
that missiles whose lifespan has expired (orange blobs)
will simply stop moving.