Work on passing renderables between containers
The Get Aux Descriptor method used by RuPPAT from
the Player class had a broken reference in it. This
was fixed.

Also, a lot of work towards being able to seamlessly
get large amounts of auxillary objects and pixels
out of an object efficiently; things such as pixels
or other objects, all related to the Object (or object
derived from Object). For example, the Player fires a
missile, the missile explodes, producing smaller fragments
made up of new Objects and pixels. These need to be
returned to RuPPAT for rendering.