Render path improved for Auc; Explosions
The render path for Objects auxillary has been
expanded to support a few more features. The
main feature at this point is the ability to
GameDestroy (a virtual, overloadable functions)
an auxillary Object, in this case a missile (orange
blob). The test GameDestroy for the Missile object
is a particle explosion, which is supported by a
few new features in the Sprite class; mainly the
addition of a vector of Pixel_desc created when the
sprite is loaded. This makes it easy to quickly grab
a vector of the pixels that make up the sprite.
Which is great for pixel based explosions.

Current issues in the Pixel color and its format. The
orange blob is made up of, you guessed it, orange pixels,
but the pixels grabbed and used in the explosion are
blue. This is likely due to formatting and data types not being
preserved somewhere.