Diagnosing color problems, implement quick fix
There seems to be issues related to the endianess when grabbing pixel
color values from a loaded sprite and placing them as single pixels in
the usual manner. It seems that the endianess has been swapped and thus
reverses the order of the color from RRGGBBAA to AABBGGRR. The quick fix
is just a reversal of this order within the Renderer. A more elegant fix
is needed as well as a better understanding as to why this is actually
happening.