Added Common static class
The new statis class "Common" (all static members) will hold generic
functions that are needed in various areas of the engine. These are
mostly very specific functions, functions that do fall into a certain
class and do not really share much. The first functions implemented here
are the ApplyDimming method and the TestBounds method. These functions
are relatively simple and are needed in several places. This simplifies
a lot of code by avoiding duplicate functions in serparate classes
and should make the engine more easily expanded.