Refactoring; Organization; Missile Destruction
RuPPAT.cpp was reorganized in a more sensible manner.
Several classes had useless things removed and the methods
for handling object destruction and pixel/sprite passing
was streamlined. This concluded with the completion of the basic
missile behavior: you shoot a missile, it explodes into particles,
all relvant back end data is destroyed.