TextDriver begining implementation
Relatively significant additions, in the from of a new class and new
consolidations in Common, to bring simple and effective text regions.
Currently, there is a lot of difficulty getting the alpha data to blit to
a temporary surface, followed by bliting the temporary surface to the
main screen. Rather than fight with it now, I'm going to start my
transition to SDL2.0, this way, any headway I make won't have to be
reimplemented in the API. And knows, maybe apla blending is snap in the
new SDL (it's really just that I can't wrap my head around multi-layer
alpha transfer)