Functional port to SDL2.0 COMPLETE!
Result is better performance through use of the SDL_texture (even for
the particles). Screen will also scale, but beware, not all objects
respond to this scaling correctly. Also, alpha blending is no longer a
complete crap-shoot with multiple surface blits. This relieves a lot of
headaches.

Major changes had to be made to accommodate the main thread restrictions
of popular graphics back ends (OpenGL). This means that only the main
thread is able to access key graphics object like the renderer. This is
the primary motivation for the Render Job queue, which is still in
early implementation, but working.

The primitives class has not been ported. Depending on primitive
support, I may remove this class.