Radial slicing now used for collisions
The object class now uses the silhouette of the sprite (2d vector of
booleans) to determine the average distance the edge of the sprite is
from the center per "slice". For instance, if there are four slices
(relatively low granularity), then each slice is 90 degrees. The class
then samples each slice area to determine the average radius in that
area. These radii are then used to check collisions by correlating
the angle between the objects, which then indicates the slice radius to
use.

The code for this is still really ugly, I'm still getting some
exception in some places, and the slice to use is not offset by the
rotation of the object...yet!